Refactor: 게임 소켓을 커스텀 훅으로 변경 #64
